following intruction ...
With our new "routeEmployee(id)" method in place, we can now update our EmployeesComponent template
(employees.component.html) to invoke the method by binding a "click" event to the same <tr> element that
has our *ngFor directive. This click event will invoke the routeEmployee() method and pass in the value of the
current employee's _id value, ie "routeEmployee(employee._id)"
By adding the click binding to the repeated <tr> element, we can ensure that every single <tr> element will
have the correct event, and invoke the routeEmployee method for the matching employee _id value for that
row in the table.